% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multidiff.R
\name{multi_compare}
\alias{multi_compare}
\title{Compares Samples using different regression methods.}
\usage{
multi_compare(
  df,
  benchmark,
  independent,
  dependent,
  method,
  rm_na = "pairwise",
  out_output_list = T,
  out_df = T,
  print_p = F,
  print_se = F,
  weight = NULL,
  id = NULL,
  strata = NULL,
  nest = FALSE,
  weight_bench = NULL,
  id_bench = NULL,
  strata_bench = NULL,
  nest_bench = FALSE,
  robust_se = F,
  p_adjust = NULL,
  names_df_benchmark = NULL,
  silence_summary = F
)
}
\arguments{
\item{df, benchmark}{A Data frame containing the sample or benchmark to
compare, or a character string containing the name of the sample or
benchmark. All independent and dependent variables must be inside both data
frames.}

\item{independent}{A list of strings containing the independent variables (x) for comparison.
Every independent variable will be used in every model to estimate the dependent variable (y)}

\item{dependent}{A list of strings containing the dependent variables (y) for comparison.
One model will be computed for every dependent variable (y) provided.}

\item{method}{A character string for the method used for comparison. It can either be "ols"
(Default), or "logit".}

\item{rm_na}{A character to determine how to handle missing values.  For this two
options are  supportet. If \code{rm_na = "pairwise"} NAs will be removed
seperately for every model. Only cases containing NA on one of the variables used
in the respective model will be removed (all independent variables but only
the respective dependent variable). If \code{rm_na = "listwise"} all cases containing NA
on one of the dependent or independent variables are removed.}

\item{out_output_list}{A logical value. If \code{out_output_list} = TRUE, a list will be
returned, containing the seperate interaction models calculated with the glm function (
or svyglm in case of weighting, as well as a summary object for every model.
Standard errors and p-values of this models are always calculated without robustness methods.}

\item{out_df}{IF True, the used data frames will also be part of the output list}

\item{print_p}{IF TRUE, in addition to the difference in Average Discrete Change (ADC),
p-values will be printed.}

\item{print_se}{IF TRUE, additionally Standard Errors will be printed.}

\item{weight, weight_bench}{A character vector containing the name of the weight
variable in the respective data frame. IF provided the data frame will be weighted
using the \code{survey package}. Also id must be provided.}

\item{id, id_bench}{A character vector containing the name of the id variable in the respectiv
data frame. Only needed for weighting.}

\item{strata, strata_bench}{A character vector containing the name of the strata variable
in the respective data frame. It is used in the \code{\link[survey]{svydesign}}
function for weighting.}

\item{nest, nest_bench}{A logical Vector used in the \code{\link[survey]{svydesign}}
function for the respective data frame.}

\item{robust_se}{A logical value If TRUE instead of normal standard errors,
heteroscedasticity-consistent standard errors will be used in the analysis for
calculation the sandwitch package and lmtest packages are used.}

\item{p_adjust}{A logical input or character string indicating a adjustment method usable in the}

\item{names_df_benchmark}{A vector containing first the name of df and benchmark.}

\item{silence_summary}{A Logical value, to indicate if the
printed summary should not be printed instead.

\code{method} parameter of \code{\link[stats]{p.adjust}}. If set to TRUE the Bonferroni adjusted
p-values are used in inference.
statistic.}
}
\value{
A table is printed showing the difference between samples for each model
, as well as an indicator, if they differ significantly from each other. It is
generated using the chosen \code{method}. If\code{out_output_list} = TRUE, also a list
with additional information will be returned that can be used in some additional
packages of this function to reprint the summary or to visualize the results.
}
\description{
\code{multi_compare} compares samples using regression models based on differing methods.
For now, only OLS-Regression and Logit-Regression are implemented.
}
\examples{

## Get Data for comparison
card<-wooldridge::card

south <- card[card$south==1,]
north <- card[card$south==0,]

## use the function to plot the data 
multi_data1<-sampcompR::multi_compare(df = north, 
                                     bench = south,
                                     independent = c("age","fatheduc","motheduc","IQ"),
                                     dependent = c("educ","wage"),
                                     method = "ols") 
                        

}
