% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multidiff_2.R
\name{multi_compare2}
\alias{multi_compare2}
\title{Compares Samples using any glm or svyglm.}
\usage{
multi_compare2(
  df,
  benchmark,
  formula_list,
  rm_na = "pairwise",
  out_output_list = T,
  out_df = T,
  print_p = F,
  print_se = F,
  weight = NULL,
  id = NULL,
  strata = NULL,
  nest = FALSE,
  weight_bench = NULL,
  id_bench = NULL,
  strata_bench = NULL,
  nest_bench = FALSE,
  robust_se = F,
  p_adjust = NULL,
  names_df_benchmark = NULL,
  family = stats::gaussian(link = "identity"),
  silence_summary = F,
  ...
)
}
\arguments{
\item{df, benchmark}{Data frames containing the samples to compare. All independent and
dependent variables must be inside both data frames.}

\item{formula_list}{A list of named formulas. Every formula in the list will
be given to \code{\link[stats]{glm}} or \code{\link[survey]{svyglm}}.}

\item{rm_na}{A character to determine how to handle missing values.  For this two
options are  supportet. If \code{rm_na = "pairwise"} NAs will be removed
seperately for every model. Only cases containing NA on one of the variables used
in the respective model will be removed (all independent variables but only
the respective dependent variable). If \code{rm_na = "listwise"} all cases containing NA
on one of the dependent or independent variables are removed.}

\item{out_output_list}{A logical value. If \code{out_output_list} = TRUE, a list will be
returned, containing the seperate interaction models calculated with the glm function (
or svyglm in case of weighting, as well as a summary object for every model.
Standard errors and p-values of this models are always calculated without robustness methods.}

\item{out_df}{IF True, the used data frames will also be part of the output list}

\item{print_p}{IF TRUE, in addition to the difference in Average Discrete Change (ADC),
p-values will be printed.}

\item{print_se}{IF TRUE, additionally Standard Errors will be printed.}

\item{weight, weight_bench}{A character vector containing the name of the weight
variable in the respective data frame. IF provided the data frame will be weighted
using the \code{survey package}. Also id must be provided.}

\item{id, id_bench}{A character vector containing the name of the id variable in the respectiv
data frame. Only needed for weighting.}

\item{strata, strata_bench}{A character vector containing the name of the strata variable
in the respective data frame. It is used in the \code{\link[survey]{svydesign}}
function for weighting.}

\item{nest, nest_bench}{A logical Vector used in the \code{\link[survey]{svydesign}}
function for the respective data frame.}

\item{robust_se}{A logical value If TRUE instead of normal standard errors,
heteroscedasticity-consistent standard errors will be used in the analysis for
calculation the sandwitch package and lmtest packages are used.}

\item{p_adjust}{A logical input or character string indicating a adjustment method usable in the
\code{method} parameter of \code{\link[stats]{p.adjust}}. If set to TRUE the Bonferroni adjusted
p-values are used in inference.}

\item{names_df_benchmark}{A vector containing first the name of df and benchmark.}

\item{family}{A family input, that can be given to \code{\link[stats]{glm}} or
\code{\link[survey]{svyglm}}}

\item{silence_summary}{A Logical value, to indicate if the
printed summary should not be printed instead.}

\item{...}{Potential other paramaters that can be given to \code{\link[stats]{glm}} or

\code{\link[survey]{svyglm}}.}
}
\value{
A table is pinted showing the difference between samples for each model
, as well as an indicator, if they differ sinificantly from each other. It is
generated using the chosen \code{method}. If\code{out_output_list} = TRUE, also a list
with additional informations will be retruned that can be used in some aditional
packages of this function to reprint the summary or to visualize the results.
}
\description{
\code{multi_compare2} compares samples using regression models based on glm other
than \link{multi_compare},it is able to compare all types of \code{\link[stats]{glm}}
or \code{\link[survey]{svyglm}} models.
}
\examples{

## Get Data for comparison
card<-wooldridge::card

south <- card[card$south==1,]
north <- card[card$south==0,]

formula_list<-list(as.formula(educ ~ age + fatheduc + motheduc + IQ),
                   as.formula(wage ~ age + fatheduc + motheduc + age*IQ))
names(formula_list)<-c("Education", "Wage")

## use the function to plot the data 
multi_data1<-sampcompR::multi_compare2(df = north, benchmark= south,
                                       formula_list=formula_list)
 
 #Visualize the difference                                       
 sampcompR::plot_multi_compare("multi_data1")
                        

}
