% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document_the_sampcompR_package.R
\name{sampcompR}
\alias{sampcompR}
\title{sampcompR: A package for the comparison of samples}
\description{
The sampcompR package provides a bundle of functions, to easily analyze and
visualise differences between Samples or Subsamples on three levels. The
comparisons can be univariate, bivariate or multivariate. Visualization is
based on \code{\link[ggplot2:ggplot]{ggplot}}
and can be edited as other plots of ggplot afterwards.
}
\section{sampcompR functions}{

\describe{
\item{\link[sampcompR]{uni_compare}}{Compare Datasets Univariate and Plot Differences}
\item{\link[sampcompR]{plot_uni_compare}}{Plot uni_compare objects}
\item{\link[sampcompR]{uni_compare_table}}{Get a table output of a uni_compare object}
\item{\link[sampcompR]{R_indicator}}{Calculate the R_indicator for several surveys}
\item{\link[sampcompR]{biv_compare}}{Compare Datasets Bivariate and Plot Differences}
\item{\link[sampcompR]{plot_biv_compare}}{Plot biv_compare objects}
\item{\link[sampcompR]{biv_compare_table}}{Get a table output of a biv_compare object}
\item{\link[sampcompR]{multi_compare}}{Compare two Datasets on a Multivariate Level (Any GLM Model)}
\item{\link[sampcompR]{plot_multi_compare}}{Plot multi_compare objects}
\item{\link[sampcompR]{multi_compare_table}}{Get a table output of multi_compare objects}
\item{\link[sampcompR]{multi_compare_merge}}{Combine two multi_compare objects, to plot them together}
\item{\link[sampcompR]{descriptive_table}}{Get a Descriptive Table for Every Data Frame}
\item{\link[sampcompR]{dataequalizer}}{Equalize dataframes}
}
}

\section{uni_compare function}{

\link[sampcompR]{uni_compare} Returns data or a plot showing the difference of two or more
data frames The differences are calculated on the base of
differing metrics, chosen in the funct argument.
Results can be visualized using \code{\link[sampcompR]{plot_uni_compare}}.
}

\section{biv_compare function}{

\link[sampcompR]{biv_compare} Returns data or heatmap of difference between two or
more data frames, by comparing their correlation matrices. The comparison is
based on Pearson's r, calculated using the \code{\link[Hmisc]{rcorr}} function.
Results can be visualized using \code{\link[sampcompR]{plot_biv_compare}}.
}

\section{multi_compare function}{

\link[sampcompR]{multi_compare} Returns data of difference between two data frames
on a multivariate level. Similar (multivariate) regression models are
compared between the surveys. Only GLM models are possible. Results can be
visualized using \code{\link[sampcompR]{plot_multi_compare}}.
}

\section{dataequalizer function}{

\link[sampcompR]{dataequalizer} compares two data frames and looks if data frames contain columns
with the same name. A copy of y is returned, containing only columns named identical
in x and y data frames. The function is mainly used in the other functions of the package.

_PACKAGE
}

